+++ Snowflake.Data/Core/Session/SessionPool.cs
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) 2012-2023 Snowflake Computing Inc. All rights reserved.
+ */
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,12 +13,10 @@ using Snowflake.Data.Log;
 
 namespace Snowflake.Data.Core.Session
 {
-    sealed class SessionPoolSingleton : IDisposable
+    sealed class SessionPool : IDisposable
     {
-        private static readonly SFLogger s_logger = SFLoggerFactory.GetLogger<SessionPoolSingleton>();
-        private static SessionPoolSingleton s_instance = null;
+        private static readonly SFLogger s_logger = SFLoggerFactory.GetLogger<SessionPool>();
         private static readonly object s_sessionPoolLock = new object();
-
         private readonly List<SFSession> _sessionPool;
         private int _maxPoolSize;
         private long _timeout;
@@ -22,7 +24,7 @@ namespace Snowflake.Data.Core.Session
         private const long Timeout = 3600;
         private bool _pooling = true;
 
-        SessionPoolSingleton()
+        internal SessionPool()
         {
             lock (s_sessionPoolLock)
             {
@@ -31,7 +33,8 @@ namespace Snowflake.Data.Core.Session
                 _timeout = Timeout;
             }
         }
-        ~SessionPoolSingleton()
+        
+        ~SessionPool()
         {
             ClearAllPools();
         }
@@ -39,21 +42,6 @@ namespace Snowflake.Data.Core.Session
         public void Dispose()
         {
             ClearAllPools();
-        }
-
-        public static SessionPoolSingleton Instance
-        {
-            get
-            {
-                lock (s_sessionPoolLock)
-                {
-                    if(s_instance == null)
-                    {
-                        s_instance = new SessionPoolSingleton();
-                    }
-                    return s_instance;
-                }
-            }
         }
 
         private void CleanExpiredSessions()
@@ -246,5 +234,4 @@ namespace Snowflake.Data.Core.Session
             return _pooling;
         }
     }
-
 }
